CHANGE: Renombrar columnas de timestamps a nomenclatura en español y actualizar funciones relacionadas

+++ backend/identity/src/main/java/com/docflow/identity/application/ports/UserWithRolesProjection.java
@@ -23,7 +23,7 @@ public record UserWithRolesProjection(
     String email,
     String nombreCompleto,
     String estado,
-    Integer rolId,
+    Long rolId,
     String rolCodigo,
     String rolNombre,
     OffsetDateTime fechaCreacion
